refactor: move Offset unit to src/Offset subdirectory
